requirements.txt
tweepy==4.14.0
pytest==8.3.4
requests==2.32.3
oauthlib==3.2.2
requests-oauthlib==1.3.1
fastapi>=0.115.2
uvicorn==0.27.0
python-dotenv>=1.0.1
jinja2>=3.1.4
websockets==12.0
anthropic==0.8.1
openai>=1.55.3
litellm>=1.55.10
pillow>=11.0.0
python-multipart>=0.0.18
psutil==5.9.8
duckduckgo-search==7.2.1
smolagents>=1.1.0 replicate
